# Slim down Completion and ExpressionResult struct sizes