--- a/Jint/Engine.Modules.cs
+++ b/Jint/Engine.Modules.cs
@@ -149,7 +149,7 @@ namespace Jint
             }
             else if (promise.State == PromiseState.Rejected)
             {
-                ExceptionHelper.ThrowJavaScriptException(this, promise.Value, new Completion(CompletionType.Throw, promise.Value, null, Location.From(new Position(), new Position(), specifier)));
+                ExceptionHelper.ThrowJavaScriptException(this, promise.Value, new Completion(CompletionType.Throw, promise.Value, null, Completion._emptyNode));
             }
             else if (promise.State != PromiseState.Fulfilled)
             {