--- a/Jint/Runtime/Interpreter/EvaluationContext.cs
+++ b/Jint/Runtime/Interpreter/EvaluationContext.cs
@@ -19,7 +19,7 @@ namespace Jint.Runtime.Interpreter
         public Completion ResumedCompletion { get; }
         public bool DebugMode { get; }
 
-        public Node LastSyntaxNode { get; set; } = null!;
+        public SyntaxElement LastSyntaxElement { get; set; } = null!;
         public bool OperatorOverloadingAllowed { get; }
     }
 }