--- a/Jint/Runtime/Interpreter/Expressions/JintUnaryExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintUnaryExpression.cs
@@ -46,7 +46,7 @@ namespace Jint.Runtime.Interpreter.Expressions
             // need to notify correct node when taking shortcut
             context.LastSyntaxNode = _expression;
 
-            return Completion.Normal(EvaluateJsValue(context), _expression.Location);
+            return Completion.Normal(EvaluateJsValue(context), _expression);
         }
 
         protected override ExpressionResult EvaluateInternal(EvaluationContext context)