--- a/Jint/Runtime/Interpreter/JintStatementList.cs
+++ b/Jint/Runtime/Interpreter/JintStatementList.cs
@@ -78,7 +78,7 @@ namespace Jint.Runtime.Interpreter
             }
 
             JintStatement? s = null;
-            var c = new Completion(CompletionType.Normal, null!, null, context.LastSyntaxNode?.Location ?? default);
+            var c = new Completion(CompletionType.Normal, null!, null, context.LastSyntaxNode);
             Completion sl = c;
 
             // The value of a StatementList is the value of the last value-producing item in the StatementList
@@ -96,7 +96,7 @@ namespace Jint.Runtime.Interpreter
                             c.Type,
                             c.Value ?? sl.Value,
                             c.Target,
-                            c.Location);
+                            c._source);
                     }
                     sl = c;
                     lastValue = c.Value ?? lastValue;
@@ -104,8 +104,7 @@ namespace Jint.Runtime.Interpreter
             }
             catch (JavaScriptException v)
             {
-                var location = v.Location == default ? s!.Location : v.Location;
-                var completion = new Completion(CompletionType.Throw, v.Error, null, location);
+                var completion = new Completion(CompletionType.Throw, v.Error, null, s!._statement);
                 return completion;
             }
             catch (TypeErrorException e)
@@ -114,7 +113,7 @@ namespace Jint.Runtime.Interpreter
                 {
                     e.Message
                 });
-                return new Completion(CompletionType.Throw, error, null, s!.Location);
+                return new Completion(CompletionType.Throw, error, null, s!._statement);
             }
             catch (RangeErrorException e)
             {
@@ -122,9 +121,9 @@ namespace Jint.Runtime.Interpreter
                 {
                     e.Message
                 });
-                c = new Completion(CompletionType.Throw, error, null, s!.Location);
+                c = new Completion(CompletionType.Throw, error, null, s!._statement);
             }
-            return new Completion(c.Type, lastValue ?? JsValue.Undefined, c.Target, c.Location);
+            return new Completion(c.Type, lastValue ?? JsValue.Undefined, c.Target, c._source);
         }
 
         /// <summary>