Both `Completion` and `EvaluationResult` were needlessly using full copy of `Location` for error reporting needs. It's enough to hold on to source Esprima syntax element and then get the location from there when needed. This improves stack capacity greatly.

Relates to #1159

/cc @KurtGokhan @CreepGin 

## Old layouts

```
Type layout for 'Completion'
Size: 48 bytes. Paddings: 4 bytes (%8 of empty space)
|=======================================|
|   0-7: JsValue Value (8 bytes)        |
|---------------------------------------|
|  8-15: String Target (8 bytes)        |
|---------------------------------------|
| 16-19: CompletionType Type (4 bytes)  |
| |================================|    |
| |   0-3: Int32 value__ (4 bytes) |    |
| |================================|    |
|---------------------------------------|
| 20-23: padding (4 bytes)              |
|---------------------------------------|
| 24-47: Location Location (24 bytes)   |
| |===================================| |
| |   0-7: String Source (8 bytes)    | |
| |-----------------------------------| |
| |  8-15: Position Start (8 bytes)   | |
| | |===============================| | |
| | |   0-3: Int32 Line (4 bytes)   | | |
| | |-------------------------------| | |
| | |   4-7: Int32 Column (4 bytes) | | |
| | |===============================| | |
| |-----------------------------------| |
| | 16-23: Position End (8 bytes)     | |
| | |===============================| | |
| | |   0-3: Int32 Line (4 bytes)   | | |
| | |-------------------------------| | |
| | |   4-7: Int32 Column (4 bytes) | | |
| | |===============================| | |
| |===================================| |
|=======================================|


Type layout for 'ExpressionResult'
Size: 40 bytes. Paddings: 7 bytes (%17 of empty space)
|===============================================|
|   0-7: Object Value (8 bytes)                 |
|-----------------------------------------------|
|     8: ExpressionCompletionType Type (1 byte) |
| |==============================|              |
| |     0: Byte value__ (1 byte) |              |
| |==============================|              |
|-----------------------------------------------|
|  9-15: padding (7 bytes)                      |
|-----------------------------------------------|
| 16-39: Location Location (24 bytes)           |
| |===================================|         |
| |   0-7: String Source (8 bytes)    |         |
| |-----------------------------------|         |
| |  8-15: Position Start (8 bytes)   |         |
| | |===============================| |         |
| | |   0-3: Int32 Line (4 bytes)   | |         |
| | |-------------------------------| |         |
| | |   4-7: Int32 Column (4 bytes) | |         |
| | |===============================| |         |
| |-----------------------------------|         |
| | 16-23: Position End (8 bytes)     |         |
| | |===============================| |         |
| | |   0-3: Int32 Line (4 bytes)   | |         |
| | |-------------------------------| |         |
| | |   4-7: Int32 Column (4 bytes) | |         |
| | |===============================| |         |
| |===================================|         |
|===============================================|
```

## New layouts

```
Type layout for 'Completion'
Size: 32 bytes. Paddings: 7 bytes (%21 of empty space)
|========================================|
|   0-7: SyntaxElement _source (8 bytes) |
|----------------------------------------|
|  8-15: JsValue Value (8 bytes)         |
|----------------------------------------|
| 16-23: String Target (8 bytes)         |
|----------------------------------------|
|    24: CompletionType Type (1 byte)    |
| |==============================|       |
| |     0: Byte value__ (1 byte) |       |
| |==============================|       |
|----------------------------------------|
| 25-31: padding (7 bytes)               |
|========================================|
Type layout for 'ExpressionResult'
Size: 24 bytes. Paddings: 7 bytes (%29 of empty space)
|===============================================|
|   0-7: SyntaxElement _source (8 bytes)        |
|-----------------------------------------------|
|  8-15: Object Value (8 bytes)                 |
|-----------------------------------------------|
|    16: ExpressionCompletionType Type (1 byte) |
| |==============================|              |
| |     0: Byte value__ (1 byte) |              |
| |==============================|              |
|-----------------------------------------------|
| 17-23: padding (7 bytes)                      |
|===============================================|
```