@lahma 2022-08-27
## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|242.89 ms|0.522 ms|3000.0000|-|-|51,129 KB|
| **New** |	|	| **221.63 ms (-9%)** | **0.270 ms** | **3000.0000 (0%)** | **333.3333** | **-** | **51,125 KB (0%)** |
| Old |Run|3d-morph|185.44 ms|0.450 ms|3000.0000|1000.0000|333.3333|49,744 KB|
| **New** |	|	| **164.95 ms (-11%)** | **0.861 ms** | **2000.0000 (-33%)** | **1000.0000 (0%)** | **-** | **49,744 KB (0%)** |
| Old |Run|3d-raytrace|198.13 ms|0.354 ms|5000.0000|1000.0000|-|93,394 KB|
| **New** |	|	| **172.07 ms (-13%)** | **2.265 ms** | **5000.0000 (0%)** | **1000.0000 (0%)** | **-** | **93,390 KB (0%)** |
| Old |Run|access-binary-trees|106.03 ms|0.413 ms|4000.0000|1000.0000|-|78,472 KB|
| **New** |	|	| **88.14 ms (-17%)** | **0.220 ms** | **4000.0000 (0%)** | **1000.0000 (0%)** | **-** | **78,472 KB (0%)** |
| Old |Run|access-fannkuch|577.46 ms|9.080 ms|-|-|-|137 KB|
| **New** |	|	| **492.58 ms (-15%)** | **1.327 ms** | **-** | **-** | **-** | **136 KB (-1%)** |
| Old |Run|access-nbody|227.20 ms|0.482 ms|3666.6667|333.3333|-|63,851 KB|
| **New** |	|	| **201.69 ms (-11%)** | **0.946 ms** | **3666.6667 (0%)** | **333.3333 (0%)** | **-** | **63,850 KB (0%)** |
| Old |Run|access-nsieve|187.28 ms|0.781 ms|1000.0000|-|-|21,519 KB|
| **New** |	|	| **161.81 ms (-14%)** | **0.302 ms** | **1000.0000 (0%)** | **-** | **-** | **21,518 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|200.59 ms|0.811 ms|4000.0000|-|-|68,950 KB|
| **New** |	|	| **165.83 ms (-17%)** | **0.261 ms** | **4000.0000 (0%)** | **-** | **-** | **68,950 KB (0%)** |
| Old |Run|bitops-bits-in-byte|312.72 ms|0.970 ms|2500.0000|-|-|45,446 KB|
| **New** |	|	| **255.65 ms (-18%)** | **0.552 ms** | **2000.0000 (-20%)** | **-** | **-** | **45,446 KB (0%)** |
| Old |Run|bitops-bitwise-and|220.48 ms|0.751 ms|5000.0000|-|-|93,443 KB|
| **New** |	|	| **179.70 ms (-18%)** | **1.107 ms** | **5000.0000 (0%)** | **-** | **-** | **93,439 KB (0%)** |
| Old |Run|bitops-nsieve-bits|243.86 ms|0.714 ms|3000.0000|1000.0000|-|54,856 KB|
| **New** |	|	| **216.25 ms (-11%)** | **0.581 ms** | **3000.0000 (0%)** | **1000.0000 (0%)** | **-** | **54,856 KB (0%)** |
| Old |Run|contr(...)rsive [21]|139.83 ms|0.381 ms|7500.0000|1750.0000|-|124,120 KB|
| **New** |	|	| **123.34 ms (-12%)** | **0.240 ms** | **7400.0000 (-1%)** | **1600.0000 (-9%)** | **-** | **124,117 KB (0%)** |
| Old |Run|crypto-aes|156.47 ms|0.401 ms|-|-|-|15,605 KB|
| **New** |	|	| **139.10 ms (-11%)** | **0.412 ms** | **-** | **-** | **-** | **15,590 KB (0%)** |
| Old |Run|crypto-md5|124.87 ms|0.236 ms|5000.0000|1000.0000|-|95,416 KB|
| **New** |	|	| **108.00 ms (-14%)** | **0.295 ms** | **5000.0000 (0%)** | **1000.0000 (0%)** | **-** | **95,414 KB (0%)** |
| Old |Run|crypto-sha1|130.26 ms|0.342 ms|5000.0000|1000.0000|-|82,075 KB|
| **New** |	|	| **115.77 ms (-11%)** | **0.424 ms** | **5000.0000 (0%)** | **1000.0000 (0%)** | **-** | **82,074 KB (0%)** |
| Old |Run|date-format-tofte|121.79 ms|0.314 ms|2000.0000|-|-|48,636 KB|
| **New** |	|	| **107.11 ms (-12%)** | **0.179 ms** | **2000.0000 (0%)** | **-** | **-** | **48,549 KB (0%)** |
| Old |Run|date-format-xparb|66.00 ms|0.157 ms|1750.0000|750.0000|-|29,415 KB|
| **New** |	|	| **60.27 ms (-9%)** | **0.166 ms** | **1777.7778 (+2%)** | **666.6667 (-11%)** | **-** | **29,415 KB (0%)** |
| Old |Run|math-cordic|430.43 ms|2.908 ms|7000.0000|-|-|123,974 KB|
| **New** |	|	| **358.22 ms (-17%)** | **1.057 ms** | **7000.0000 (0%)** | **-** | **-** | **123,973 KB (0%)** |
| Old |Run|math-partial-sums|149.78 ms|0.548 ms|4000.0000|-|-|68,947 KB|
| **New** |	|	| **135.03 ms (-10%)** | **0.383 ms** | **4000.0000 (0%)** | **-** | **-** | **68,947 KB (0%)** |
| Old |Run|math-spectral-norm|154.12 ms|0.389 ms|4000.0000|-|-|67,176 KB|
| **New** |	|	| **139.22 ms (-10%)** | **0.469 ms** | **4000.0000 (0%)** | **-** | **-** | **67,176 KB (0%)** |
| Old |Run|regexp-dna|106.98 ms|0.474 ms|800.0000|800.0000|800.0000|17,671 KB|
| **New** |	|	| **105.89 ms (-1%)** | **0.487 ms** | **1000.0000 (+25%)** | **1000.0000 (+25%)** | **1000.0000 (+25%)** | **17,670 KB (0%)** |
| Old |Run|string-base64|120.23 ms|0.648 ms|-|-|-|12,527 KB|
| **New** |	|	| **108.52 ms (-10%)** | **0.236 ms** | **-** | **-** | **-** | **12,526 KB (0%)** |
| Old |Run|string-fasta|224.79 ms|0.520 ms|10000.0000|333.3333|-|167,276 KB|
| **New** |	|	| **205.58 ms (-9%)** | **0.625 ms** | **10000.0000 (0%)** | **333.3333 (0%)** | **-** | **167,275 KB (0%)** |
| Old |Run|string-tagcloud|83.20 ms|0.292 ms|2000.0000|1000.0000|-|48,438 KB|
| **New** |	|	| **76.39 ms (-8%)** | **0.267 ms** | **2000.0000 (0%)** | **1000.0000 (0%)** | **-** | **48,438 KB (0%)** |
| Old |Run|string-unpack-code|83.25 ms|0.343 ms|5000.0000|500.0000|-|84,575 KB|
| **New** |	|	| **76.65 ms (-8%)** | **0.242 ms** | **5000.0000 (0%)** | **571.4286 (+14%)** | **-** | **84,573 KB (0%)** |
| Old |Run|strin(...)input [21]|1,223.99 ms|24.086 ms|742000.0000|738000.0000|737000.0000|6,352,229 KB|
| **New** |	|	| **1,252.85 ms (+2%)** | **24.291 ms** | **703000.0000 (-5%)** | **699000.0000 (-5%)** | **698000.0000 (-5%)** | **6,351,929 KB (0%)** |