@KurtGokhan 2022-08-27
@lahma thank you for this. Results look good. 

I don't understand what the table in the above comment should be showing though. Shouldn't `Allocated` be lower in the new runs?