@lahma 2022-08-27
Nope, all allocations happen on stack (reducing  recursion capability as it's recursion × stack frame size). Benchmarkdotnet doesn't show value types in memory allocations.